+++ src/dashboard/apigateway/apigateway/apis/open/esb/permission/serializers.py
@@ -31,7 +31,7 @@ from apigateway.apps.permission.constants import (
     PermissionApplyExpireDaysEnum,
     PermissionStatusEnum,
 )
-from apigateway.biz.slz_validators import BKAppCodeValidator
+from apigateway.biz.validators import BKAppCodeValidator
 from apigateway.common.fields import TimestampField
 from apigateway.utils import time
