+++ src/dashboard/apigateway/apigateway/apis/open/permission/serializers.py
@@ -32,7 +32,7 @@ from apigateway.apps.permission.constants import (
 )
 from apigateway.apps.permission.models import AppPermissionRecord
 from apigateway.biz.permission import PermissionDimensionManager
-from apigateway.biz.slz_validators import BKAppCodeValidator, ResourceIDValidator
+from apigateway.biz.validators import BKAppCodeValidator, ResourceIDValidator
 from apigateway.common.fields import TimestampField
 from apigateway.utils import time
