+++ src/dashboard/apigateway/apigateway/apps/resource/serializers.py
@@ -32,7 +32,7 @@ from apigateway.apps.stage.serializers import HostSLZ, TransformHeadersSLZ, Upst
 from apigateway.apps.support.resource_doc.utils import get_resource_doc_key
 from apigateway.biz.resource import ResourceHandler
 from apigateway.biz.resource_url import ResourceURLHandler
-from apigateway.biz.slz_validators import MaxCountPerGatewayValidator
+from apigateway.biz.validators import MaxCountPerGatewayValidator
 from apigateway.common.contexts import ResourceAuthContext
 from apigateway.common.fields import CurrentGatewayDefault
 from apigateway.common.mixins.serializers import ExtensibleFieldMixin
